[table] Merge MultiRowDataTable and DataTable, then replace DataTable with MultiRowDataTable and move the original to legacy as DataTableOld, including it's dependencies.
